Change suspect code completion test
This change tests that, in string interpolation, code completion doesn’t suggest Void functions but suggests functions returning other types.

This is a change in behavior, but I’m not convinced the old version was correct in the first place—why would you want to interpolate a function that returns Void? It’s valid but not useful, and code completion seems to try to avoid suggesting Void functions in other similar circumstances.